projects/gemini/pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.30.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
